Create SQL_Database_Girl_Scout_Troop
This code is repeatable and contains both DDL and DML language. I created a database for the Girl Scout Troop I lead. This version has names, phone numbers, and email addresses changed for privacy. On the database that I maintain and is set to private, the information is accurate so as to be useful. This version also is not updated with newer events and badges.